Using Recursive loop
The main idea of solution is to use as much multiplications as possible, for example how can we evaluate x^20? We can just multiply x in loop 20 times, but we also can evaluate x^10 and multiply it by itself! Similarly, x^10 = x^5 * x^5. Now we have odd power, but it is not a problem, we evaluate x^5 = x^2 * x^2 * x. We also need to deal with some border cases, here is the full algorithm:

If we have very small value of x we can directly return 0, the smallest value of float is 1.175494 × 10^(-38).
If we have n = 0, return 1.
If we have negative power, return positive power of 1/x.
Now, we have two cases: for even and for odd n, where we evaluate power n//2.
Complexity: time complexity is O(log n), space complexity for this recursive algorithm is also O(log n), which can be reduced to O(1), if we use iterative approach instead.