Must return the same array in which the elements were given as input
Here, when we choose the first element, we find its final position after rotation. Then, we replace the element at that position with the first one and the replaces element is again stored in a temporary variable. It's final position is again calculated and the process is repeated. Suppose an already replaced position is obtained, then the next element to that position is taken and similar operation is done.