By making a DP array
The array created contains the whole grid in which the robot will move. So when traversing each cell of the array, we add the number of ways to reach that particular cell based on the number given from the previous or adjacent (only top and left cells connected to the current one) cells. Pretty easy if you get used to doing DP question.