Simple logic of two pointers
If the number is greater than the target number, then answer is the difference between the number and the target. Otherwise, if the number is lesser than the target, first check if the target number is odd or even. If odd, then increment the count and increment the target value by 1, to make it even. Then, if the given number is greater than or equal to half of the target number, then the the difference between the number and the target's half plus one will be added to the count. Stop the looping right there. Otherwise, change the target value to its half and increment the count by 1, which means that one step is required, that is doubling the half value. Then continue with the loop.