Greedy Algorithm used here
Here, when an element is pushed into the stack, then inside every iteration of for loop, a while loop is executed, which checks if the top of the stack is equal to the current element pointing in the 'popped' array. If this condition is met, then the top of the stack is popped and the pointer in the 'popped' array is incremented by one. This operation is carried out until either the stack is empty or the pointer has reached the end of the 'popped' array. Then, after execution of all the iterations in the for loop, then it is checked if the pointer of the 'popped' array has reached the end or not. If it has reached, that means all the specified elements on the 'popped' array has successfully been popped, in the given order, thus, returning True, else False.