Using Recursion (DID IT ON MY OWN)
In this question, there are two ways of moving when reached at a number in an array. Either go i-arr[i] or i+arr[i]. This step repeats recursively at each element of the array. Thus, the extreme conditions of the recursive function would be that if the index exceeds the size of the array (or is lesser than 0th index), then return False. There can also be a repetitive condition where the function keeps looping in the same indices. So, a visited array is also made so as to keep track of all the visited indices. Thus, if that particular index is already visited , then return False again. Otherwise, if arr[i] == 0, then return True. Since, there are two conditions of moving in the array, the results from the recursive calls are stored in the two variables 'r' and 't'. Thus, the "OR" operation of the two variables is returned, so as to maintain the "True" condition, if obtained in any of the recursive calls.