Using two pointer approach
First, the left pointer is made to search for the first '1' in the array. When found, it can also be noted that the maximum required distance is also equal to that left pointer value. This is because, in case arr[0] is not equal to one, then the maximum distance from the nearest '1' will be just the left pointer's value. So, the right pointer is right next to the left pointer (towards its right). When iterated, if the right pointer lands on a 'one', then the maximum distance is taken as athe maximum values between current max and the mid point between right and left pointers. thus, the right pointer is assigned to the left pointer and the right pointer is incremented by one. Finally, after iteration, if the rightmost element of the array is '0', then the maximum distance is the maximum between the current max and the distance between the right and left pointers.