USING DFS (RECURSION)
Basically, on each node, it is checked if its left and right subtree returns a node (which matches p or q). This recursion happens in the function itself which is given; no need to define a separate recursive function. If the node matches either p or q, then it is returned to its parent node. Now, from the parent, if the other subtrees are checked and the other node is not found, then it means that the other node might be a descendant of the first "found" node. Hence, return that "found" node.